remove functions from headers * Functions
